chore: update Go version and dependencies (#143)
* chore: update Go version and dependencies

Upgrade Go to version 1.25 and update various dependencies to their latest versions for improved performance and security. This includes updates to the kairos-sdk, docker, and opentelemetry packages.

Signed-off-by: Itxaka <itxaka@kairos.io>

* chore: update Go version to 1.25

This change updates the Go version in the Earthfile and unit-tests.yml to 1.25-bookworm to ensure compatibility with the latest features and improvements.

Signed-off-by: Itxaka <itxaka@kairos.io>

* go mod tidy

Signed-off-by: Itxaka <itxaka@kairos.io>

* go mod tidy

Signed-off-by: Itxaka <itxaka@kairos.io>

* Fix?

Signed-off-by: Itxaka <itxaka@kairos.io>

* Fix

Signed-off-by: Itxaka <itxaka@kairos.io>

* maybe fix

Signed-off-by: Itxaka <itxaka@kairos.io>

* Fix

Signed-off-by: Itxaka <itxaka@kairos.io>

---------

Signed-off-by: Itxaka <itxaka@kairos.io>